requirements.txt
youtube-transcript-api
google-generativeai
google-genai
python-dotenv
-e .